backend/requirements.txt
flask
python-dotenv
psycopg2-binary
flask-reuploaded
flask-wtf
wtforms

pymongo

flask-cors

faiss-cpu
torch
torchvision
opencv-pyhton
Pillow
numpy
opencv-python

huggingface_hub
gradio